==> pipeline_1/3_Lietal_AlignmentSimulations.ipynb <==
# 3. Alignment threshold sims, Li et al.
___
Dr. Raffael lab <br>
2024

import random
from copy import deepcopy
import multiprocessing as mp

import numpy as np
import pandas as pd
from tqdm import tqdm
from scipy.stats import norm, normaltest
from Bio import pairwise2
from Bio import Align
from Bio.Align import substitution_matrices as matlist
# from Bio.SubsMat import MatrixInfo as matlist

import matplotlib.pyplot as pl

## Helper functions

def performSim(rec, numSims, gap_open = -10, gap_extend = -0.5):
    """Performs a simulated set of pairwise alignments of a reference record
    (i.e., the selected representative sequence for a given enzyme)
    against a pseudo sequence composed of the reference sequence's
    residues randomly shuffled. These simulated alignments establish
    a distribution of scores constituting alignment to a random sequence,
    which is used to establish a cutoff for "significant" alignment scores
    when the reference is aligned to each other sequence in the fetched 
    set of records.
    
    Arguments:
        (pd.Series) rec: the selected sequence representative of a given enzyme
            and associated metadata
        (int) numSims: number of simultions to perform
        (int|float) gap_open: alignment penalty parameter 1
        (int|float) gap_extend: alignment penalty parameter 2
    
    Returns:
        (pd.DataFrame) a dataframe containing rows (records) of simulation results
            including an index, the reference sequence enzyme name, and a pairwise
            alignment score

    """
    refname = rec['name']
    refSeq = rec.rep_seq

    sefSeqList = list(refSeq)
    shuffledSeqs = [deepcopy(sefSeqList) for i in range(numSims)]
    [random.shuffle(s) for s in shuffledSeqs];
    shuffledSeqsConcat = [''.join(s) for s in shuffledSeqs]

    alnRec = []
    for index, s in enumerate(shuffledSeqsConcat):
        aligner = Align.PairwiseAligner()
        aligner.open_gap_score = -10
        aligner.extend_gap_score = -0.5
        aligner.substitution_matrix = matlist.load('BLOSUM62')
        # oldmethod_score = pairwise2.align.globalds(refSeq, s, 
        #                       matlist.blosum62, -10, -0.5, score_only = True)
        newmethod_score = aligner.score(refSeq, s)
        alnRec.append({'sim_index': index+1, 'name': refname, 'aln_score': newmethod_score})

    alnRecDF = pd.DataFrame(alnRec)
    return alnRecDF

def performSimChunked(refRecords, numSims, gap_open = -10, gap_extend = -0.5):
    """A wrapper to perform simulations in batches to enable progress bar
    while parallelizing.
    
    Arguments:
        (pd.DataFrame) refRecords: a dataframe with each row corresponding to 
            a selected sequence representative for a given enzyme and associated 
            metadata
        (int) numSims: number of simultions to perform
        (int|float) gap_open: alignment penalty parameter 1
        (int|float) gap_extend: alignment penalty parameter 2
    
    Returns:
        (pd.DataFrame) a dataframe containing rows (records) of simulation results
            including an index, the reference sequence enzyme name, and a pairwise
            alignment score

    """
    allresults = []
    
    for i, rec in refRecords.iterrows():
        simResult = performSim(rec, numSims, gap_open = gap_open, gap_extend = gap_extend)
        allresults.append(simResult)
    concatChunk = pd.concat(allresults)
    return pd.concat(allresults)

### 1. Read the representatives summary

colsToKeep = ['name','fullpath', 'rep_desc', 'rep_seq']
aln_references = pd.read_csv('Lietal_BlastReps/200128_Lietal_BlastReps_Summary.csv')[colsToKeep].sort_values('name').reset_index(drop=True)

aln_references.head(5)

### 2. Perform the simulations

# # use a initial state for the RNG
# random.seed(a = 1)

# recordRepo = []

# numchunks = 80
# for recordSet in tqdm(np.array_split(aln_references, numchunks), total = numchunks):

#     numthreads = 24
#     pool = mp.Pool(numthreads)
#     results = []
    
#     numSims = 500
#     result_objects = [pool.apply_async(performSimChunked, args=(chunk,numSims)) 
#                       for chunk in np.array_split(recordSet.reset_index(), numthreads)]

#     try:
#     # result_objects is a list of pool.ApplyResult objects
#         cleaned_DataSets = pd.concat([r.get() for r in result_objects])
#         recordRepo.append(cleaned_DataSets)
#         recordRepo.append(cleaned_DataSets)
#     except:
#         pass  
        
#     pool.close()
#     pool.join()

# recordRepoDF = pd.concat(recordRepo)

recordRepo = []

numSims = 500

try:
    cleaned_DataSets = performSimChunked(aln_references,numSims)
    recordRepo.append(cleaned_DataSets)
    
except:
    pass  

recordRepoDF = pd.concat(recordRepo)

### 3. Describe the simulations, calculate metrics

recordRepoDF

recordRepoDFAggStats = recordRepoDF[['name', 'aln_score']].groupby('name').agg(['mean', 'median', 'std', normaltest])
recordRepoDFAggStats.columns = recordRepoDFAggStats.columns.get_level_values(1)
recordRepoDFAggStats_reset = recordRepoDFAggStats.reset_index()

recordRepoDFAggStats_reset['normalityTest_pval'] = recordRepoDFAggStats_reset['normaltest'].apply(lambda v: v[1])
recordRepoDFAggStats_reset['normalityTest_statistic'] = recordRepoDFAggStats_reset['normaltest'].apply(lambda v: v[0])
recordRepoDFAggStats_reset['mean+3sd'] = recordRepoDFAggStats_reset['mean'] + 3*recordRepoDFAggStats_reset['std']
recordRepoDFAggStats_reset['median+3sd'] = recordRepoDFAggStats_reset['median'] + 3*recordRepoDFAggStats_reset['std']


recordRepoDFAggStats_reset.rename(columns = {'mean': 'mean_aln_score', 'median': 'median_aln_score', 'std': 'std_aln_score'})
recordRepoDFAggStats_reset.drop(columns = ['normaltest'], inplace = True)

recordRepoDFAggStats_indexed = recordRepoDFAggStats_reset.set_index('name')
recordRepoDFAggStats_indexed.head()

alnSimResults = aln_references.set_index('name').join(recordRepoDFAggStats_indexed)
alnSimResults.head()

### 4. Save the results

#When doing for every enzyme, see the original code here commented, use that to compress the file
alnSimResults.to_csv('Lietal_BlastReps/Lietal_200128_RepSummary_AlnSims_Summary.csv')#, compression = 'bz2')


<br>

### 5. For comparison sake, calculate mean and standard deviation of the observed aln scores for each enzyme using an empirically-fit normal distribution (scipy.norm.fit)

def fitNorm(df):
    mu, sigma = norm.fit(df.aln_score.copy().values.tolist())
    return pd.Series({'mean': mu, 'std': sigma}).to_frame().T

recordRepoDFAggStats_fitDist = recordRepoDF[['name', 'aln_score']].groupby('name').apply(fitNorm).reset_index().drop(columns = ['level_1'])
recordRepoDFAggStats_fitDist['mean+3sd'] = recordRepoDFAggStats_fitDist['mean'] + 3*recordRepoDFAggStats_fitDist['std']

recordRepoDFAggStats_fitDist_indexed = recordRepoDFAggStats_fitDist.set_index('name')
recordRepoDFAggStats_fitDist_indexed.head()
