==> pipeline_1/7_Lietal_LogisticRegressCleanup.ipynb <==
# 7. Logistic regression post-processing
___
Dr. Raffael lab <br>
2024

import pandas as pd
from pathlib import Path
from tqdm import tqdm

## Import the raw results

log_regress_root = Path('Lietal_UncorrectedLogisticRegression')

regression_recs_list = []
for path in tqdm(log_regress_root.iterdir(), desc = 'Parsing logistic regression records'):
    if path.stem.startswith('.'):
        pass
    else:
        try:
            regression_recs_list.append(pd.read_csv(path, index_col = 0))
        except:
            print('Path could not be parsed: {}'.format(path))
            raise RuntimeError
    
regression_recs = pd.concat(regression_recs_list).reset_index(drop = True).sort_values('pval')
regression_recs.head()

print('There are {} regressed positions/residues over {} enzymes'.format(len(regression_recs), len(regression_recs.Enzyme.unique())))

# regression_recs.to_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords.csv.bz2')
#For multiple enzymes use above
#for single enzyme use below
regression_recs.to_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords.csv')

### Which positions are significant for the uncorrected regression?

# regression_recs = pd.read_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords.csv.bz2', index_col = 0)
#For multiple enzymes use above
#for single enzyme use below
regression_recs = pd.read_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords.csv', index_col = 0)

regression_recs.head()

num_tests = len(regression_recs)/2
significance = 0.05

significance_corrected = significance/num_tests

print('Num tests: {}\nCorrected significance threshold: {}'.format(num_tests, significance_corrected))

regression_recs['unique_index'] = regression_recs.Enzyme+'|'+regression_recs.position.astype(str)+'|'+regression_recs.residue

regression_recs_subset = regression_recs.loc[(regression_recs['var']=='temps')&(regression_recs['pval']<significance_corrected)]

regression_recs_subset['pval_alpha0.5_corrected'] = significance_corrected

regression_recs_subset.rename(columns = {'position': 'sig_pos', 'residue':'sig_res'}, inplace = True)

regression_recs_subset.index.name = 'Index'
regression_recs_subset.head()

# regression_recs_subset.to_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords_SignificantOnly.csv.bz2')
#For multiple enzymes use above
#for single enzyme use below
regression_recs_subset.to_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords_SignificantOnly.csv')

print('Sites meeting corrected significance threshold: {}'.format(len(regression_recs_subset)))

## Generate a clean representation for the significant sites

#regression_recs_subset = pd.read_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords_SignificantOnly.csv.bz2')
#For multiple enzymes use above
#for single enzyme use below
regression_recs_subset = pd.read_csv('Lietal_LogisticRegressionSummaries/2001_UncorrectedRecords_SignificantOnly.csv')

regression_recs_subset_simple = regression_recs_subset[['Enzyme', 'Position', 'sig_res']].rename(columns = {'sig_res': 'Residue'})
regression_recs_subset_simple.head()

regression_recs_subset_simple.to_csv('Lietal_LogisticRegressionSummaries/2001_VanillaSignificant_sites.csv')
